requirements.txt
astroid==3.1.0
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.2.0
certifi==2024.2.2
click==8.1.7
dill==0.3.8
docstring-to-markdown==0.15
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
jedi==0.19.1
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pylint==3.1.0
PySocks==1.7.1
pytest==8.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
tomlkit==0.12.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.1
wsproto==1.2.0
